More line filter tests (upper/lower sideband position)
