+++ tests/base/test_base.py
@@ -3,9 +3,7 @@ import pytest
 from libertem.utils.devices import detect
 
 from libertem_holo.base.utils import HoloParams
-from libertem_holo.base.generate import hologram_frame
 from libertem_holo.base.reconstr import get_phase, phase_offset_correction
-from libertem_holo.base.filters import butterworth_disk, butterworth_line
 
 
 @pytest.mark.with_numba
@@ -82,74 +80,6 @@ def test_params_sb_upper_lower(holo_data):
     )
     assert p1.sb_position_int == (53, 58)
     assert p2.sb_position_int == (11, 6)
-
-
-def test_butterworth_disk_cpu_gpu_equiv():
-    d = detect()
-    if not d['cudas'] or not d['has_cupy']:
-        pytest.skip("No CUDA device or no CuPy, skipping CuPy test")
-
-    import cupy as cp
-    disk_gpu = butterworth_disk(shape=(512, 511), radius=128.0, order=12, xp=cp)
-    disk_cpu = butterworth_disk(shape=(512, 511), radius=128.0, order=12, xp=np)
-
-    assert np.allclose(disk_gpu.get(), disk_cpu)
-
-
-def test_butterworth_line_cpu_gpu_equiv():
-    d = detect()
-    if not d['cudas'] or not d['has_cupy']:
-        pytest.skip("No CUDA device or no CuPy, skipping CuPy test")
-
-    import cupy as cp
-
-    line_cpu = butterworth_line(
-        shape=(512, 511),
-        width=3,
-        sb_position=(100.1, 100),
-        length_ratio=0.9,
-        order=12,
-        xp=np,
-    )
-    line_gpu = butterworth_line(
-        shape=(512, 511),
-        width=3,
-        sb_position=(100.1, 100),
-        length_ratio=0.9,
-        order=12,
-        xp=cp,
-    )
-
-    assert np.allclose(line_gpu.get(), line_cpu)
-
-
-def test_linefilter_orientation_l_r():
-    sy, sx = 64, 64
-    ref = hologram_frame(np.ones((sy, sx)), np.zeros((sy, sx)), f_angle=30)
-    ref_rot = hologram_frame(np.ones((sy, sx)), np.zeros((sy, sx)), f_angle=120)
-    p = HoloParams.from_hologram(ref)
-    p_rot = HoloParams.from_hologram(ref_rot)
-    line = butterworth_line(
-        shape=(64, 64),
-        width=10,
-        sb_position=p.sb_position_int,
-        order=25
-    )
-    line_rot = butterworth_line(
-        shape=(64, 64),
-        width=10,
-        sb_position=p_rot.sb_position_int,
-        order=25
-    )
-
-    # line filter comes from top-left:
-    assert np.allclose(line[0, 0], 0)
-
-    # bottom right is all-ones:
-    assert np.allclose(line[-1, -1], 1)
-
-    # in the rotated case, line filter comes from bottom left:
-    assert np.allclose(line_rot[-1, 0], 0)
 
 
 @pytest.mark.with_numba
